St_Exupery_de_Rust

Function world_set_scenario

pub(crate) fn world_set_scenario(
    scene: Scenario,
    monsters: &mut Vec<Monstre>,
    hero: &mut StExupery,
) -> bool

‘Match’ le ‘Scenario’ produit par world_get_scenario() à un scénario parmi trouver une potion, une arme ou un bouclier ou, de combattre un monstre et exécute le scénario.

#Arguments

‘scene’, ‘monsters’, ‘hero’

#Retourne

bool (‘true’ si points de vie de l’instance StExupery <= 0)

Exemples

‘’’ let hero_state = world_set_scenario(scene, monsters, hero); ‘’’