St_Exupery_de_Rust

Module models

Structs

Bouclier

Bouclier modélise une arme défensive qui est un attribut d’une instance StExupery.

Epee

Épée modélise une arme qui est un attribut d’une instance StExupery.

Loot

Loot modélise un entrepôt d’instances uniques Epee et Bouclier.

Monstre

Monstre modélise l’adversaire principal du jeu et ses comportements.

Potion

Potion modélise un élément de l’attribut vector ‘inventaire’ d’une instance StExupery. Cet item permet au joueur de rehausser la valeur de l’attribut ‘vie’ d’une instance StExupery.

StExupery

St_Exupery modélise le joueur et ses actions avec un struct composé.