St_Exupery_de_Rust::models

Struct StExupery

pub struct StExupery {
    vie: i16,
    pub epee: Epee,
    pub bouclier: Bouclier,
    pub inventaire: Vec<Potion>,
}

St_Exupery modélise le joueur et ses actions avec un struct composé.

Attributs:

vie (i16): représente la vitalité du joueur.

epee (struct Epee): modèle Epee pour modéliser l’arme du joueur.

bouclier (struct Bouclier): modèle Bouclier modélise une arme pour réduire les dommages recus.

inventaire

Exemples

‘’’ let h = StExupery::new(&Diff::F); ‘’’

Fields
vie: i16epee: Epeebouclier: Bouclierinventaire: Vec<Potion>
Implementations

impl StExupery

pub fn new(diff: &Diff) -> Self

Initialize une nouvelle instance ‘StExupery’ selon le niveau de difficulté choisit par l’utilisateur.

#Arguments

‘diff’ - niveau de difficulté

#Retourne

‘Self’

Exemples

‘’’ let h = StExupery::new(diff);

impl StExupery

pub fn get_health(&self) -> i16

Retourne le niveau de vie d’une instance StExupery

#Arguments

Aucuns

#Retourne

‘self.vie’

Exemples

‘’’ let h = StExupery::default(); println!(“{}”, h.get_health()) ‘’’

impl StExupery

pub fn set_health(&mut self, valeur: i16)

Modifie la valeur de l’attribut ‘vie’ d’une instance StExupery

#Arguments

‘valeur’

#Retourne

( )

Exemples

‘’’ let h = StExupery::default(); h.set_health(45)

impl StExupery

pub fn get_shield(&mut self) -> &mut Bouclier

Retourne la valeur de l’attribut ‘bouclier’ d’une instance StExupery

#Arguments

Aucuns

#Retourne

‘&mut Bouclier’

Exemples

‘’’ let h = StExupery::default(); println!(“{}”,h.get_shield()); ‘’’

impl StExupery

pub fn get_sword(&mut self) -> &mut Epee

Retourne la valeur de l’attribut ‘epee’ d’une instance StExupery

#Arguments

Aucuns

#Retourne

‘&mut Epee’

Exemples

‘’’ let h = StExupery::default(); println!(“{}”,h.get_sword()); ‘’’

impl StExupery

pub fn get_inventory(&self) -> usize

Retourne le nombre de potions en inventaire d’une instance StExupery

#Arguments

Aucuns

#Retourne

‘usize’

Exemples

‘’’ let h = StExupery::default(); println!(“{}”,h.get_inventory()); ‘’’

impl StExupery

pub fn ajouter_potion(&mut self)

Tente d’ajouter une potion dans l’inventaire d’une instance StExupery

#Arguments

Aucuns

#Retourne

( )

Exemples

‘’’ let h = StExupery.default(); h.ajouter_potion(); ‘’’

impl StExupery

pub fn boire_potion(&mut self)

Tente de boire une potion de l’inventaire et indique la valeur de ‘vie’ d’une instance StExupery. #Arguments

Aucuns

#Retourne

( )

Exemples

‘’’ let h = StExupery.default(); h.boire_potion(); ‘’’

impl StExupery

pub fn boire_potion_trouvee(&mut self)

Tente de boire une potion trouvée lors de la navigation du monde de jeu et indique la valeur de ‘vie’ d’une instance StExupery apr la suite.

#Arguments

Aucuns

#Retourne

( )

Exemples

‘’’ let h = StExupery.default(); h.boire_potion_trouvee(); ‘’’

impl StExupery

pub fn take_damage(&mut self, dam: i16) -> bool

Pour une instance StExupery: recoit la valeur de dommages infligés par un monstre, ajuste cette valeur en fonction de la valeur d’un attribut de ‘bouclier’ et réduit la valeur de ‘vie’ en conséquence. Évalue si ‘vie’ <= 0 et retourne ‘true’ pour indiqué le “décès” de l’instance.

#Arguments

‘dam’

#Retourne

bool

Exemples

‘’’ let h = StExupery.default(); h.take_damage(6) ‘’’

impl StExupery

pub fn attack(&mut self) -> i16

Additionne les points de vie d’une instance StExupery divisés par 10 et l’attribut de puissance de l’attribut ‘epee’ et retourne cette valeur.

#Arguments

*Aucuns

#Retourne

i16

Exemples

‘’’ let h = StExupery.default(); h.take_damage(6) ‘’’

impl StExupery

pub fn equip_combat_item(&mut self, item: CombatItem)

Détermine le type d’attribut (épée ou bouclier) à modifier pour une instance StExupery et met à jour l’attribut suite choix de l’utilisateur d’équipper l’item.

#Arguments

‘item’

#Retourne

( )

Exemples

‘’’ let h = StExupery.default(); h.equip_combat_item(CombatItem::Sword(de_type_epee)) ‘’’